Add files via upload
Each of the scenario consist of five cam’s coverage which are covering whole environment setup. All cam is synchronized with frames starting and ending point. Target track with different angles to make resolution difference.To limit the movement options of UAV’s, make environment as a quad-tree-discretized The environment is fixed, and the number of targets use multiple in our case. No obstacle present in the environment setup. Size of the environment is 488 cm length and 172 cm width. Ground_Truth cam or parent cam cover whole environment and rest of four cam each cover ¼ coordinates of the environment. micro-aerial vehicles (MAVs) Coordinate sensor movement for searching a target is the main issue in robotic vision. each MAV takes sensor observation of target existence and updates search region local informations.